Adds a item abstration to Memoizable and Callable Items * Adds a Factory to create an Item * Adds Memoizable and Callable classes to avoid some conditions
fix travis

try again
