+++ lib/dry/container.rb
@@ -1,7 +1,10 @@
 require 'concurrent'
 require 'dry-configurable'
 require 'dry/container/error'
-require 'dry/container/item'
+require 'dry/container/item/base'
+require 'dry/container/item/memoizable'
+require 'dry/container/item/callable'
+require 'dry/container/item/factory'
 require 'dry/container/namespace'
 require 'dry/container/registry'
 require 'dry/container/resolver'
